requirements.txt
python-dotenv>=0.9.9
langchain>=1.0.2
langchain-community>=0.4.1
langchain-core>=1.0.1
langchain-openai>=1.0.1
langgraph>=1.0.1
google-adk>=0.1.0
passlib>=1.7.4
pydantic-settings>=2.11.0
pymongo>=4.14.1
streamlit>=1.50.0
argon2-cffi>=23.1.0
Pillow>=10.0.0
huggingface_hub>=0.20.0
requests>=2.31.0
